Toujours plus de corrections

+++ ft_strlcat.c
@@ -14,14 +14,14 @@
 
 size_t	ft_strlcat(char *dst, const char *src, size_t size)
 {
-	int	i;
-	int	j;
+	size_t	i;
+	size_t	j;
 
 	i = 0;
 	j = 0;
 	while (dst[i])
 		i++;
-	while (src[j] && i + j + 1 < (int)size)
+	while (src[j] && i + j + 1 < size)
 	{
 		dst[i + j] = src[j];
 		j++;
@@ -30,7 +30,7 @@ size_t	ft_strlcat(char *dst, const char *src, size_t size)
 	j = 0;
 	while (src[j])
 		j++;
-	if ((int)size <= i)
+	if (size <= i)
 		return (size + j);
 	return (i + j);
 }
